Tagging version 2.2
git-svn-id: https://plugins.svn.wordpress.org/purify-wp-menues/tags/2.2@963445 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ uninstall.php
@@ -19,12 +19,24 @@ if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
 * Run before deleting the plugin
 *
 * @since   1.0
-* @uses    $settings_db_slug
-* @uses    $wpdb
 */
 // remove settings
-delete_option( 'purify_wp_menu_options_set' ); 
-// clean DB
-global $wpdb;
-$wpdb->query( "OPTIMIZE TABLE `" .$wpdb->options. "`" );
+if ( is_multisite() ) {
+
+	$sites = wp_get_sites();
+
+	if ( empty ( $sites ) ) return;
+
+	foreach ( $sites as $site ) {
+		// switch to next blog
+		switch_to_blog( $site[ 'blog_id' ] );
+		// remove settings
+		delete_option( 'purify_wp_menu_options_set' );
+	}
+	// restore the current blog, after calling switch_to_blog()
+	restore_current_blog();
+} else {
+	// remove settings
+	delete_option( 'purify_wp_menu_options_set' );
+}
